Fixed issues with main class.
The main class wasn't being generated. This started because the
MANIFEST file wasn't being created. Fixed this by adding
maven-jar-plugin to the pom.xml. The next issue was that the main class
couldn't be found. This was fixed by adding the package information to
the top of each java file in the program.